+++ radio.ino
@@ -151,7 +151,7 @@ int SendRadioTelemetry(void)
 void lowSpeedTelemetry(void)
 {
       unsigned long timer=0;
-      unsigned int  GPS_Alt_tmp;
+      long  GPS_Alt_tmp;
       GetRadioHousekeeping();
       _Serial.print(F("OBC: COM temp: "));
       _Serial.println(radio_temp);
@@ -174,7 +174,7 @@ void lowSpeedTelemetry(void)
         _Serial.print(GPS_long);
         _Serial.print(F(","));
         //_Serial.print(GPS_Alt_ch);
-        _Serial.printf("%05d",GPS_Alt_tmp);
+        _Serial.printf("%05ld",GPS_Alt_tmp);
         _Serial.print(F(","));
         _Serial.printf("%04d",(int)(ext_temp*10.0));
         _Serial.print(F(","));
